unifying console display options in one place
